poc/requirements.txt
google-cloud-documentai>=2.24.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.42.0
google-auth>=2.27.0
pillow>=10.2.0
python-dotenv>=1.0.0
